LayoutTests/accessibility/mac/treeitem-row-delegation-expected.txt

This tests that a treeitem with a group that uses aria-owns will report its disclosed rows correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tree.childrenCount is 7
PASS treeitem1.isEqual(accessibilityController.accessibleElementById('item-1')) is true
PASS treeitem1_1.isEqual(accessibilityController.accessibleElementById('group1-item1')) is true
PASS treeitem1_2.isEqual(accessibilityController.accessibleElementById('group1-item2')) is true
PASS treeitem1_3.isEqual(accessibilityController.accessibleElementById('group1-item3')) is true
PASS treeitem1_2_1.isEqual(accessibilityController.accessibleElementById('group2-item1')) is true
PASS treeitem1_2_2.isEqual(accessibilityController.accessibleElementById('group2-item2')) is true
PASS treeitem2.isEqual(accessibilityController.accessibleElementById('item-2')) is true
PASS successfullyParsed is true

TEST COMPLETE
Jaguar
Jungle Animals
Birds
Reptiles
Cats
Lion
Domestic Animals
